refactor: move Utility to another file
